Simplified conditional IR nodes by replacing them with a single Predicate IR value.
